[WIP] purge: Migrate from a setuid program to a nginx perl handler
The main motivation for this change is, that this means, no compiler
is needed on the server.
We could keep the setuid program for people that prefer that.
But that's probably not worth the maintainance costs..
(and it can be checked out from git, if ever needed)

A note on why we use perl: The lua nginx module is not part of
NGINX core. Perl is part of the core system and (at least on Fedora)
installable without manual compiling.

This solution allows the nginx cache server to be located on a
different host than the TYPO3 system.